pyproject.toml
[tool.poetry]
name = "pythontools"
version = "0.1.0"
description = "Python 常用开发工具类"
authors = ["pancx <p_lord@163.com>"]
readme = "README.md"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]  # https://docs.astral.sh/ruff/settings/#lint
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP",  # pyupgrade
    "FURB",
]
ignore = [
    "E402",  # module level import not at top of file
    "W191",  # indentation contains tabs
    "B008", # do not perform function calls in argument defaults
    "C901", # function is too complex
    "E501", # line too long
]

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
pytz = "^2024.1"
cfgv = "^3.4.0"

[project]
requires-python = ">=3.11"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
types-pytz = "^2024.1.0.20240417"
ruff = "^0.5.6"
coverage = "^7.6.1"
pytest-cov = "^5.0.0"
refurb = "^2.0.0"
faker = "^27.0.0"
allure-pytest = "^2.13.5"
pre-commit = "^4.0.1"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
